useDispatch hook to call reducers or functions from store
The function interacts with the store by calling specific actions, which perform operations based on the current state. You can pass values to these functions, and these values will be captured as action.payload within the store. The function also takes a state parameter, which contains the current values from the store. Once an action is dispatched (typically via useDispatch), it triggers the reducer, which modifies the state accordingly.